Interpolation, Filter and ACF memory usage optimized

+++ lib/filtro_FFT.py
@@ -41,7 +41,7 @@ def loadRasterImage(path):
         return raster_ds, raster_ds.GetRasterBand(1).ReadAsArray(), False, ""
     else:
         rt = raster_ds
-        return raster_ds, np.stack([raster_ds.GetRasterBand(i).ReadAsArray() for i in range(1, raster_ds.RasterCount+1)], axis=2), False, ""
+        return raster_ds, np.stack([raster_ds.GetRasterBand(i).ReadAsArray() for i in range(1, raster_ds.RasterCount+1)], axis=2).astype(np.int16), False, ""
      
 def saveSingleBand(dst, rt, img, tt=gdal.GDT_Int16, typ='GTiff'): ##
     """
@@ -140,7 +140,7 @@ def getFiltRaster(path):
         sys.exit(1)
 
     # Auxiliar
-    aux = np.zeros(img.shape)
+    aux = np.zeros(img.shape).astype(np.int16)
 
     # Dims
     height, width, depth = img.shape
@@ -154,9 +154,9 @@ def getFiltRaster(path):
     start = True
     for i, j in itertools.product(range(height), range(width)):
             #aux[i, j, :] = np.abs(filterFFt1d(img[i, j, :], threshold=0.05)
-            tt = np.abs(filterFFt1d((np.array(img[i, j, :])), threshold=0.05))
+            tt = np.abs(filterFFt1d((np.array(img[i, j, :])), threshold=0.05)).astype(np.int16)
             #tt = np.abs(filterFFt1d(max_filter1d_valid(np.array(img[i, j, :]), 3), threshold=0.05))
-            aux[i, j, :] = np.pad(tt, (0, depth - len(tt)), 'constant')
+            aux[i, j, :] = np.pad(tt, (0, depth - len(tt)), 'constant').astype(np.int16)
             rmse[i, j] = np.sqrt(np.sum(np.power(img[i, j, :] - aux[i, j, :], 2))/depth)##
             pearson[i, j] = st.pearsonr(img[i, j, :], aux[i, j, :])[0]##  
             progress = int((i*width + j)/(height*width) * 100)
